third_party/blink/web_tests/external/wpt/webrtc/historical-expected.txt

This is a testharness.js-based test.
[FAIL] RTCDataChannel member reliable should not exist
  assert_false: expected false got true
[FAIL] RTCPeerConnection member addStream should not exist
  assert_false: expected false got true
[FAIL] RTCPeerConnection member createDTMFSender should not exist
  assert_false: expected false got true
[FAIL] RTCPeerConnection member getLocalStreams should not exist
  assert_false: expected false got true
[FAIL] RTCPeerConnection member getRemoteStreams should not exist
  assert_false: expected false got true
[FAIL] RTCPeerConnection member onaddstream should not exist
  assert_false: expected false got true
[FAIL] RTCPeerConnection member onremovestream should not exist
  assert_false: expected false got true
[FAIL] RTCPeerConnection member removeStream should not exist
  assert_false: expected false got true
[FAIL] webkitRTCPeerConnection interface should not exist
  assert_false: expected false got true
Harness: the test ran to completion.
